Use MariaDB FTS index for contact searches.
This is a crude attempt at improving the UX when searching for contacts in large
deployments. It requires an index to be set up as this:

    CREATE FULLTEXT INDEX oc_cards_properties_name_FTS ON nextcloud.oc_cards_properties (value);

It is *not* comaptible with PostgreSQL yet as the FTS support there is vastly
different and more complex to handel than the MariaDB one.